Better, sorted file selection
We now sort files alphabetically while putting directories at the top.
Also, the file view scrolling now behaves like the pattern view, in that
we scroll down as soon as the cursor reaches the middle of the screen,
stopping whenever all items fit on screen.

In addition, if loading a file fails, an error message is printed on the
bottom screen with further instructions.